refactor:#192 api 함수 폴더 분리
