+++ src/app/(auth)/mypage/my-gonggam-post-list/page.tsx
@@ -1,5 +1,5 @@
 import MyGonggamPostClient from '@/components/features/mypage/my-gonggam-post-client';
-import { getUserProfile } from '@/lib/utils/api/auth.api';
+import { getUserProfile } from '@/lib/utils/api/auth/auth.api';
 
 const MyGonggamPostListPage = async () => {
   //유저 닉네임 조회
