+++ src/app/(auth)/mypage/page.tsx
@@ -1,5 +1,5 @@
 import EditProfileForm from '@/components/features/mypage/edit-profile-form';
-import { getUserProfile } from '@/lib/utils/api/auth.api';
+import { getUserProfile } from '@/lib/utils/api/auth/auth.api';
 
 const MyPage = async () => {
   const profile = await getUserProfile();
