+++ src/app/(life)/life/page.tsx
@@ -2,7 +2,7 @@
 import { redirect } from 'next/navigation';
 import LifePageClient from '@/components/features/life/life-page-client';
 import { PATH } from '@/constants/page-path';
-import { getUserProfile } from '@/lib/utils/api/auth.api';
+import { getUserProfile } from '@/lib/utils/api/auth/auth.api';
 
 const LifePage = async () => {
   const profile = await getUserProfile();
