+++ src/app/api/auth/profile/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from 'next/server';
 import { FAIL } from '@/constants/messages';
-import { getUserProfile } from '@/lib/utils/api/auth.api';
+import { getUserProfile } from '@/lib/utils/api/auth/auth.api';
 
 /**
  * 클라이언트에서 getUserProfile() 함수를 이용해서 사용자 프로필 정보를 받아오기 위한 라우트 핸들러입니다.
