+++ src/components/features/home/main-gonggam-preview.tsx
@@ -2,7 +2,7 @@ import GonggamItem from '@/components/features/home/main-gonggam-item';
 import GonggamPreviewContainer from '@/components/features/home/main-gonggam-preview-container';
 import ListContainer from '@/components/features/home/main-gonggam-preview-list-container';
 import TitleSection from '@/components/features/home/main-gonggam-preview-title-section';
-import { getGonggamPreviewList } from '@/lib/utils/api/gonggam-board.api';
+import { getGonggamPreviewList } from '@/lib/utils/api/gonggam/gonggam-board.api';
 
 const GonggamPreviewSection = async () => {
   try {
