+++ src/components/features/home/main-random-mission.tsx
@@ -2,7 +2,7 @@ import MissionButton from '@/components/features/home/main-random-mission-button
 import RandomMissionContainer from '@/components/features/home/main-random-mission-container';
 import ImageSection from '@/components/features/home/main-random-mission-image';
 import TitleSection from '@/components/features/home/main-random-title';
-import { getUserSessionState } from '@/lib/utils/api/auth.api';
+import { getUserSessionState } from '@/lib/utils/api/auth/auth.api';
 import { getMissionsData } from '@/lib/utils/api/missions.api';
 
 const RandomMissionSection = async () => {
