+++ src/components/features/mypage/edit-profile-form.tsx
@@ -10,7 +10,7 @@ import { FAIL, SUCCESS } from '@/constants/messages';
 import { useUpdateProfileMutate } from '@/lib/hooks/mutations/use-profile-update-mutate';
 import { useUserProfile } from '@/lib/hooks/queries/use-get-user-profile';
 import { useProfileForm } from '@/lib/hooks/use-profile-form';
-import { profileImageUpload } from '@/lib/utils/api/profile-image-upload.api';
+import { profileImageUpload } from '@/lib/utils/api/mypage/profile-image-upload.api';
 import { processedImage } from '@/lib/utils/processed-image';
 import { toastAlert } from '@/lib/utils/toast';
 import type { EditFormData, InitProfile } from '@/types/edit-profile-form';
