+++ src/lib/hooks/queries/use-get-user-profile.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 import { QUERY_KEY } from '@/constants/query-keys';
-import { fetchUserProfile } from '@/lib/utils/api/auth-client.api';
+import { fetchUserProfile } from '@/lib/utils/api/auth/auth-client.api';
 import type { TableUsers } from '@/types/supabase-const';
 
 export const useUserProfile = (initProfile: TableUsers) => {
