+++ src/lib/utils/api/life/life-by-mission.api.ts
@@ -1,5 +1,5 @@
 import { TABLE } from '@/constants/supabase-tables-name';
-import { fetchUserSessionState } from '@/lib/utils/api/auth-client.api';
+import { fetchUserSessionState } from '@/lib/utils/api/auth/auth-client.api';
 import { supabase } from '@/lib/utils/supabase/supabase-client';
 
 export const getLifePostByMissionId = async (missionId: number) => {
